Introduced Dashboard
GET /UserHome now receives the amount of all PageInstances and the
amount of the different stati.
This information shall be shown in a table on UserHome.jade. You may
call the /UserHome-controller in order to receive the structure of the
count-object, which will be printed on the console. (or print it in jade
by #{JSON.stringify(data.count)}

For more information on how to implement the Dashboard, see
doc/CR-Prozess.pptx on page 5.